Incorporating most of Adam's feedback.
