--- _includes/_code/tapbox-d/.gitignore
@@ -1,9 +1,0 @@
-.DS_Store
-.atom/
-.idea
-.packages
-.pub/
-build/
-ios/.generated/
-packages
-pubspec.lock
