--- _includes/_code/tapbox-d/main.dart
@@ -1,139 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:meta/meta.dart';
-
-void main() {
-  runApp(new MyApp());
-}
-
-// State is managed by TapboxModel, a subclass of ChangeNotifier.
-// The StatefulWidgets register as listeners.
-// TBD: Adam is going to add ListeningWidget to the framework,
-//      so it won't be needed here.
-// If something changes, TapboxModel notifies its listeners.
-
-//---------------------------- ParentWidget ----------------------------
-
-class ParentWidget extends StatefulWidget {
-  @override
-  ParentWidgetState createState() => new ParentWidgetState();
-}
-
-class ParentWidgetState extends State<ParentWidget> {
-  TapboxModel model = new TapboxModel();
-
-  @override
-  Widget build(BuildContext context) {
-    return new Column(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: [
-        new TapboxD(model: model),
-        new TapboxD(model: model),
-      ],
-    );
-  }
-}
-
-//-------------------------- ListeningWidget ---------------------------
-
-abstract class ListeningWidget extends StatefulWidget {
-  ListeningWidget({Key key, this.listenable}) : super(key: key);
-
-  Listenable listenable;
-
-  Widget build(BuildContext context);
-
-  ListeningWidgetState createState() => new ListeningWidgetState();
-}
-
-class ListeningWidgetState extends State<ListeningWidget> {
-  void initState() {
-    super.initState();
-    config.listenable.addListener(handleChanged);
-  }
-
-  void didUpdateConfig(ListeningWidget oldConfig) {
-    if (config.listenable != oldConfig.listenable) {
-      oldConfig.listenable.removeListener(handleChanged);
-      config.listenable.addListener(handleChanged);
-    }
-  }
-
-  void dispose() {
-    config.listenable.removeListener(handleChanged);
-    super.dispose();
-  }
-
-  void handleChanged() {
-    setState(() {});
-  }
-
-  Widget build(BuildContext context) => config.build(context);
-}
-
-//--------------------------- TapboxModel ----------------------------
-
-class TapboxModel extends ChangeNotifier {
-  bool get active => _active;
-  bool _active = false;
-  set active(bool newValue) {
-    assert(newValue != null);
-    if (_active == newValue) return;
-    _active = newValue;
-    notifyListeners();
-  }
-}
-
-//----------------------------- TapboxD ------------------------------
-
-class TapboxD extends ListeningWidget {
-  TapboxD({Key key, @required TapboxModel model})
-      : model = model,
-        super(key: key, listenable: model);
-
-  final TapboxModel model;
-
-  void handleTap() {
-    model.active = !model.active;
-  }
-
-  Widget build(BuildContext context) {
-    return new GestureDetector(
-      onTap: handleTap,
-      child: new Container(
-        child: new Center(
-          child: new Text(model.active ? 'Active' : 'Inactive',
-              style: new TextStyle(fontSize: 32.0, color: Colors.white)),
-        ),
-        width: 200.0,
-        height: 200.0,
-        decoration: new BoxDecoration(
-          backgroundColor:
-              model.active ? Colors.lightGreen[700] : Colors.grey[600],
-        ),
-      ),
-    );
-  }
-}
-
-//------------------------------- MyApp --------------------------------
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return new MaterialApp(
-      title: 'Flutter Demo',
-      theme: new ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: new Scaffold(
-        appBar: new AppBar(
-          title: new Text('Flutter Demo'),
-        ),
-        body: new Center(
-          child: new ParentWidget(),
-        ),
-      ),
-    );
-  }
-}
