--- _includes/_code/tapbox-d/pubspec.yaml
@@ -1,9 +1,0 @@
-name: tapbox_d
-description: Demonstrates exporting state to a model controller.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
